+++ src/components/whatsAppTemplates/FormTabs.vue
@@ -10,11 +10,11 @@
       :initialTab="initialTranslation"
       @change="handleTranslationSelection"
     >
-    <template #tab-head-new>
-      <div ref="new-translation-button">
-        {{ $t('WhatsApp.templates.new_language') }}
-      </div>
-    </template>
+      <template #tab-head-new>
+        <div ref="new-translation-button">
+          {{ $t('WhatsApp.templates.new_language') }}
+        </div>
+      </template>
       <template #tab-head-add>
         <div ref="add-translation-button" @click.stop="addTranslation">
           <unnnic-icon-svg icon="add-1" size="sm" />
@@ -134,7 +134,7 @@
         'errorUpdateTemplateTranslation',
       ]),
       templateSelectLanguages() {
-        return this.whatsAppTemplateSelectLanguages.map((item) => {
+        return this.whatsAppTemplateSelectLanguages?.map((item) => {
           const { text, value } = item;
           return {
             label: text,